add libary with small step
