+++ Elevato/ILift.cs
@@ -1,0 +1,8 @@
+﻿namespace Elevator
+{
+    public interface ILift
+    {
+        void GoTo();
+        void Request(int floor);
+    }
+}
