+++ ElevatorBrain/ElevatorBrain.csproj
@@ -6,4 +6,8 @@
     <UseWindowsForms>true</UseWindowsForms>
   </PropertyGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\Elevato\Elevato.csproj" />
+  </ItemGroup>
+
 </Project>
