+++ ElevatorBrain/Program.cs
@@ -17,7 +17,7 @@ namespace ElevatorBrain
             Application.SetHighDpiMode(HighDpiMode.SystemAware);
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
-            Application.Run(new Lift());
+            Application.Run(new Brain());
         }
     }
 }
